Lots of Updates (See Description)
Lots of updates in this commit including:

JUnit4 added to the build path, including a package for testing classes
and complete tests for the Pokemon and Attack Classes, and an incomplete
BattleManager test class

Various updates to the Pokemon and Attack Classes based on errors that
were found by the JUnit tests.

I also included a BattleRound class in this commit for simulating a
battle sequence, but I think I can modify the BattleManager class to
make this new class unnecessary.

Hope things are going well for everyone!