Added Player Movement
The game now renders a single player on the default map and the player
can move in terms of the cells on the map. So far, nothing is done in
terms of detecting certain zones like where the player can and can't go,
but I know how I can go about this.