Added Map Drawing Package
I added the package that all the map drawing utilities are in. Run the
main method in the MainNullLayout.java file and you'll get the
appropriate window to pop up. It does take a little while to load so be
patient.

I also added another map that I want to do some teleport tests with, but
right now I'm sick and need to go lie down...

Cheers!